Create README.md
I used an original code that was on Zybooks, where it shows a step-by-step tutorial on how to get the functions to work. The app uses GGC colors, a GGC building, and has a Floating Action Button with a calculator icon. I also changed the launcher icon from the traditional android icon to a pizza one. The last thing I did was to create a log statement that informs the user when their calculation is complete with the statement, “Your Calculation is Complete!”